execution: use UI abstraction for tool confirmation via run_agent_loop; inject ui from tools; remove direct PromptSession/create_confirm_session from execution paths; tests updated
